Project created Eclipse named SearchTrees.
In the project create a program named DFSearch that creates a DFS tree
using a file of edges named edges.txt and using the first integer in the
file as the root. The program should then asks the user to enter a node
id (integer).  The program should search for the node id and output the
following:

- the path of vertices, that are traversed until the node is found.
- the number of vertices traversed.
- the time elapsed for the search. 

In the project create a program named BFSearch that creates a BFS tree
using a file of edges named edges.txt and using the first integer in the
file as the root. The program should then asks the user to enter a node
id (integer).  The program should search for the node id and output the
following:

- the path of vertices, that are traversed until the node is found.
- the number of vertices traversed.
- the time elapsed for the search.